pyproject.toml
[tool.poetry]
name = "chat-sfdr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-community = "^0.0.37"
supabase = "^2.4.5"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.4"
langchain-openai = "^0.1.6"
langchain = "^0.1.17"
pypdf = "^4.2.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
